+++ Classes/Frontend/Request/Parameters.php
@@ -4,12 +4,11 @@ namespace Sys25\RnBase\Frontend\Request;
 
 use ArrayObject;
 use Sys25\RnBase\Utility\Arrays;
-use Sys25\RnBase\Utility\Typo3Classes;
 
 /***************************************************************
  * Copyright notice
  *
- * (c) 2007-2021 René Nitzsche <rene@system25.de>
+ * (c) 2007-2025 René Nitzsche <rene@system25.de>
  * All rights reserved
  *
  * This script is part of the TYPO3 project. The TYPO3 project is
@@ -132,12 +131,16 @@ class Parameters extends ArrayObject implements ParametersInterface
      * @param string $parameter Key (variable name) from GET or POST vars
      *
      * @return array|string returns the GET vars merged recursively onto the POST vars
+     *
+     * @deprecated use PSR-7 ServerRequestInterface
      */
-    public static function getPostAndGetParametersMerged($parameterName)
+    public static function getPostAndGetParametersMerged($parameter)
     {
-        $utility = Typo3Classes::getGeneralUtilityClass();
+        $postParameter = isset($_POST[$parameter]) && is_array($_POST[$parameter]) ? $_POST[$parameter] : [];
+        $getParameter = isset($_GET[$parameter]) && is_array($_GET[$parameter]) ? $_GET[$parameter] : [];
+        $mergedParameters = $getParameter;
 
-        return $utility::_GPmerged($parameterName);
+        return Arrays::mergeRecursiveWithOverrule($mergedParameters, $postParameter);
     }
 
     /**
@@ -146,6 +149,8 @@ class Parameters extends ArrayObject implements ParametersInterface
      * @param string $parameter Key (variable name) from GET or POST vars
      *
      * @return array|string returns the GET vars merged recursively onto the POST vars
+     *
+     * @deprecated use PSR-7 ServerRequestInterface
      */
     public static function getPostOrGetParameter($parameterName)
     {
@@ -153,29 +158,26 @@ class Parameters extends ArrayObject implements ParametersInterface
     }
 
     /**
-     * @see \TYPO3\CMS\Core\Utility\GeneralUtility::_GP
-     *
-     * @param string $parameter Key (variable name) from GET or POST vars
+     * @param string $var Key (variable name) from GET or POST vars
      *
      * @return array|string returns the GET vars merged recursively onto the POST vars
-     */
-    public static function _GP($parameterName)
-    {
-        $utility = Typo3Classes::getGeneralUtilityClass();
-
-        return $utility::_GP($parameterName);
-    }
-
-    /**
-     * @see \TYPO3\CMS\Core\Utility\GeneralUtility::_GETset
      *
-     * @param mixed  $inputGet
-     * @param string $key
+     * @deprecated use PSR-7 ServerRequestInterface
      */
-    public static function setGetParameter($inputGet, $key = '')
+    public static function _GP($var)
     {
-        $utility = Typo3Classes::getGeneralUtilityClass();
-        $utility::_GETset($inputGet, $key);
+        if (empty($var)) {
+            return;
+        }
+
+        $value = $_POST[$var] ?? $_GET[$var] ?? null;
+
+        // This is there for backwards-compatibility, in order to avoid NULL
+        if (isset($value) && !is_array($value)) {
+            $value = (string) $value;
+        }
+
+        return $value;
     }
 
     /**
@@ -186,12 +188,20 @@ class Parameters extends ArrayObject implements ParametersInterface
      * @see \TYPO3\CMS\Core\Utility\GeneralUtility::_GET
      *
      * @return mixed
+     *
+     * @deprecated use PSR-7 ServerRequestInterface
      */
     public static function getGetParameters($var = null)
     {
-        $utility = Typo3Classes::getGeneralUtilityClass();
+        $value = null === $var
+            ? $_GET
+            : (empty($var) ? null : ($_GET[$var] ?? null));
+        // This is there for backwards-compatibility, in order to avoid NULL
+        if (isset($value) && !is_array($value)) {
+            $value = (string) $value;
+        }
 
-        return $utility::_GET($var);
+        return $value;
     }
 
     /**
@@ -202,11 +212,17 @@ class Parameters extends ArrayObject implements ParametersInterface
      * @see \TYPO3\CMS\Core\Utility\GeneralUtility::_POST
      *
      * @return mixed
+     *
+     * @deprecated use PSR-7 ServerRequestInterface
      */
     public static function getPostParameters($var = null)
     {
-        $utility = Typo3Classes::getGeneralUtilityClass();
+        $value = null === $var ? $_POST : (empty($var) || !isset($_POST[$var]) ? null : $_POST[$var]);
+        // This is there for backwards-compatibility, in order to avoid NULL
+        if (isset($value) && !is_array($value)) {
+            $value = (string) $value;
+        }
 
-        return $utility::_POST($var);
+        return $value;
     }
 }