+++ Classes/Search/SearchBase.php
@@ -43,7 +43,7 @@ abstract class SearchBase
 {
     private static $instances = [];
 
-    private $tableMapping;
+    private $tableMapping = [];
 
     private $generic = false;
 
@@ -588,7 +588,7 @@ abstract class SearchBase
     private function _initSearch(array $options)
     {
         $this->setGeneric($options);
-        if (!is_array($this->tableMapping)) {
+        if ([] === $this->tableMapping) {
             $tableMapping = $this->getTableMappings();
             $tableMapping = is_array($tableMapping) ? $tableMapping : [];
             if ($this->isGeneric()) {